Updated to remove lpf1.f
Low pass filtering should now be done on samples *before* passing to
this demodulator - also do a 2x decimate.  In other words - do an LPF of
your choice (FIR/IIR and whatever shape) with a cutoff = to sample rate
* .25 then decimate the full sample rate set by 2.  This moves this
library closer to its goal of ONLY being a modulator/demodulator leaving
the rest to the host program to handle.